Bump to 17.0.2 (9d13099)
Changes preserved for NFCT and USBD MSC.

+++ components/ble/nrf_ble_qwr/nrf_ble_qwr.c
@@ -113,7 +113,7 @@ ret_code_t nrf_ble_qwr_value_get(nrf_ble_qwr_t * p_qwr,
     uint16_t handle     = BLE_GATT_HANDLE_INVALID;
     uint16_t val_len    = 0;
     uint16_t val_offset = 0;
-    uint16_t cur_len    = 0;
+    uint32_t cur_len    = 0;
 
     do
     {
